stupid me, response has different types of result messages

+++ src/main/java/com/github/jtendermint/websocket/jsonrpc/ErrorPayload.java
@@ -25,11 +25,11 @@ package com.github.jtendermint.websocket.jsonrpc;
 
 public class ErrorPayload {
 
-    private double code;
+    private int code;
     private String message;
     private String data;
 
-    public double getCode() {
+    public int getCode() {
         return code;
     }
